Abstract away annots search interface for use with page
