+++ src/search/background/annots-search.test.ts
@@ -2,7 +2,7 @@ import initStorageManager from '../memory-storex'
 import { StorageManager, getDb } from '..'
 import { AnnotsSearcher } from './annots-search'
 import normalize from 'src/util/encode-url-for-id'
-import { AnnotPage } from 'src/direct-linking/types'
+import { AnnotPage } from './types'
 import CustomListBackground from 'src/custom-lists/background'
 import AnnotsBg from 'src/direct-linking/background'
 import AnnotsStorage from 'src/direct-linking/background/storage'
